Complete all personalization options for page
Went through the entire page and updated the conetent
on index.html to add proper TIMBUKTU information.

On style.css, went through and trimmed fat for any
non-working properties in place. Furthermore, added
slight changes to look and feel of the website to as
far as TOP has taught me this far.

At this point I believe the layout portion
is finished and ready to have JS added to it. At this
point, no need to work on the page further until
javascript functionality is learned.

Pushing to Main, continuing TOP.